requirement.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.2
Django==3.2.13
django-trench==0.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.3.0
idna==3.3
PyJWT==2.3.0
pyotp==2.6.0
pytz==2022.1
requests==2.27.1
smsapi-client==2.6.0
sqlparse==0.4.2
twilio==7.8.2
typing_extensions==4.1.1
urllib3==1.26.9
yubico-client==1.13.0